# Release 0.66.0 performance fixes