--- a/.eslintrc.js
+++ b/.eslintrc.js
@@ -242,6 +242,7 @@ module.exports = {
             '@typescript-eslint/dot-notation': 'off',
             '@typescript-eslint/no-unnecessary-type-assertion': 'off',
             '@typescript-eslint/no-non-null-assertion': 'off',
+            'no-var': 'off',
             'camelcase': 'off'
         }
     }]