--- a/src/Scope.ts
+++ b/src/Scope.ts
@@ -25,6 +25,7 @@ import type { BscType } from './types/BscType';
 import { NamespaceType } from './types/NamespaceType';
 import { referenceTypeFactory } from './types/ReferenceType';
 import { unionTypeFactory } from './types/UnionType';
+import type { Identifier } from './lexer/Token';
 
 /**
  * Assign some few factories to the SymbolTable to prevent cyclical imports. This file seems like the most intuitive place to do the linking
@@ -179,7 +180,7 @@ export class Scope {
     public getEnumMemberFileLink(enumMemberName: string, containingNamespace?: string): FileLink<EnumStatement | EnumMemberStatement> {
         let lowerNameParts = enumMemberName?.toLowerCase()?.split('.');
         let memberName = lowerNameParts?.splice(lowerNameParts.length - 1, 1)?.[0];
-        let lowerName = lowerNameParts?.join('.').toLowerCase();
+        let lowerName = lowerNameParts?.join('.');
         const enumMap = this.getEnumMap();
 
         let enumeration = enumMap.get(
@@ -269,10 +270,10 @@ export class Scope {
             this.enumerateBrsFiles((file) => {
                 if (isBrsFile(file)) {
                     for (let cls of file.parser.references.classStatements) {
-                        const lowerClassName = cls.getName(ParseMode.BrighterScript)?.toLowerCase();
+                        const className = cls.getName(ParseMode.BrighterScript);
                         //only track classes with a defined name (i.e. exclude nameless malformed classes)
-                        if (lowerClassName) {
-                            map.set(lowerClassName, { item: cls, file: file });
+                        if (className) {
+                            map.set(className.toLowerCase(), { item: cls, file: file });
                         }
                     }
                 }
@@ -291,10 +292,10 @@ export class Scope {
             this.enumerateBrsFiles((file) => {
                 if (isBrsFile(file)) {
                     for (let iface of file.parser.references.interfaceStatements) {
-                        const lowerIfaceName = iface.getName(ParseMode.BrighterScript)?.toLowerCase();
+                        const ifaceName = iface.getName(ParseMode.BrighterScript);
                         //only track classes with a defined name (i.e. exclude nameless malformed classes)
-                        if (lowerIfaceName) {
-                            map.set(lowerIfaceName, { item: iface, file: file });
+                        if (ifaceName) {
+                            map.set(ifaceName.toLowerCase(), { item: iface, file: file });
                         }
                     }
                 }
@@ -312,10 +313,9 @@ export class Scope {
             const map = new Map<string, FileLink<EnumStatement>>();
             this.enumerateBrsFiles((file) => {
                 for (let enumStmt of file.parser.references.enumStatements) {
-                    const lowerEnumName = enumStmt.fullName.toLowerCase();
                     //only track enums with a defined name (i.e. exclude nameless malformed enums)
-                    if (lowerEnumName) {
-                        map.set(lowerEnumName, { item: enumStmt, file: file });
+                    if (enumStmt.fullName) {
+                        map.set(enumStmt.fullName.toLowerCase(), { item: enumStmt, file: file });
                     }
                 }
             });
@@ -332,10 +332,9 @@ export class Scope {
             const map = new Map<string, FileLink<ConstStatement>>();
             this.enumerateBrsFiles((file) => {
                 for (let stmt of file.parser.references.constStatements) {
-                    const lowerEnumName = stmt.fullName.toLowerCase();
                     //only track enums with a defined name (i.e. exclude nameless malformed enums)
-                    if (lowerEnumName) {
-                        map.set(lowerEnumName, { item: stmt, file: file });
+                    if (stmt.fullName) {
+                        map.set(stmt.fullName.toLowerCase(), { item: stmt, file: file });
                     }
                 }
             });
@@ -592,25 +591,41 @@ export class Scope {
         let namespaceLookup = new Map<string, NamespaceContainer>();
         this.enumerateBrsFiles((file) => {
             for (let namespaceStatement of file.parser.references.namespaceStatements) {
-                //TODO should we handle non-brighterscript?
-                let name = namespaceStatement.getName(ParseMode.BrighterScript);
-                let nameParts = name.split('.');
+                let nameParts = namespaceStatement.getNameParts();
 
-                let loopName = null;
+                let loopName: string = null;
+                let lowerLoopName: string = null;
+                let parentNameLower: string = null;
+
                 //ensure each namespace section is represented in the results
                 //(so if the namespace name is A.B.C, this will make an entry for "A", an entry for "A.B", and an entry for "A.B.C"
-                for (let part of nameParts) {
-                    loopName = loopName === null ? part : `${loopName}.${part}`;
-                    let lowerLoopName = loopName.toLowerCase();
+                for (let i = 0; i < nameParts.length; i++) {
+
+                    let part = nameParts[i];
+                    let lowerPartName = part.text.toLowerCase();
+
+                    if (i === 0) {
+                        loopName = part.text;
+                        lowerLoopName = lowerPartName;
+                    } else {
+                        parentNameLower = lowerLoopName;
+                        loopName += '.' + part.text;
+                        lowerLoopName += '.' + lowerPartName;
+                    }
+
                     if (!namespaceLookup.has(lowerLoopName)) {
                         namespaceLookup.set(lowerLoopName, {
                             file: file,
                             fullName: loopName,
+                            fullNameLower: lowerLoopName,
+                            parentNameLower: parentNameLower,
+                            nameParts: nameParts.slice(0, i),
                             nameRange: namespaceStatement.nameExpression.range,
-                            lastPartName: part,
+                            lastPartName: part.text,
+                            lastPartNameLower: lowerPartName,
                             namespaces: new Map(),
-                            classStatements: {},
-                            functionStatements: {},
+                            classStatements: new Map(),
+                            functionStatements: new Map(),
                             enumStatements: new Map(),
                             constStatements: new Map(),
                             statements: [],
@@ -618,13 +633,13 @@ export class Scope {
                         });
                     }
                 }
-                let ns = namespaceLookup.get(name.toLowerCase());
+                let ns = namespaceLookup.get(lowerLoopName);
                 ns.statements.push(...namespaceStatement.body.statements);
                 for (let statement of namespaceStatement.body.statements) {
                     if (isClassStatement(statement) && statement.name) {
-                        ns.classStatements[statement.name.text.toLowerCase()] = statement;
+                        ns.classStatements.set(statement.name.text.toLowerCase(), statement);
                     } else if (isFunctionStatement(statement) && statement.name) {
-                        ns.functionStatements[statement.name.text.toLowerCase()] = statement;
+                        ns.functionStatements.set(statement.name.text.toLowerCase(), statement);
                     } else if (isEnumStatement(statement) && statement.fullName) {
                         ns.enumStatements.set(statement.fullName.toLowerCase(), statement);
                     } else if (isConstStatement(statement) && statement.fullName) {
@@ -638,14 +653,9 @@ export class Scope {
 
             //associate child namespaces with their parents
             for (let [, ns] of namespaceLookup) {
-                let parts = ns.fullName.split('.');
-
-                if (parts.length > 1) {
-                    //remove the last part
-                    parts.pop();
-                    let parentName = parts.join('.');
-                    const parent = namespaceLookup.get(parentName.toLowerCase());
-                    parent.namespaces.set(ns.lastPartName.toLowerCase(), ns);
+                if (ns.parentNameLower) {
+                    const parent = namespaceLookup.get(ns.parentNameLower);
+                    parent.namespaces.set(ns.lastPartNameLower, ns);
                 }
             }
         });
@@ -812,9 +822,11 @@ export class Scope {
                 // or as a member to the containing namespace.
                 let previousNSType = currentNSType;
                 const nsNamePart = nsNamePartToken.text;
+
                 currentNSType = currentNSType === null
                     ? symbolTable.getSymbolType(nsNamePart, getTypeOptions)
                     : currentNSType.getMemberType(nsNamePart, getTypeOptions);
+
                 nameSoFar = nameSoFar === '' ? nsNamePart : `${nameSoFar}.${nsNamePart}`;
                 let isFinalNamespace = nameSoFar.toLowerCase() === fullNamespaceName.toLowerCase();
                 if (!isNamespaceType(currentNSType)) {
@@ -833,7 +845,7 @@ export class Scope {
                     }
                 }
 
-                // Now the namespace type is built, add the aggregate as a sibling
+                // Now that the namespace type is built, add the aggregate as a sibling
                 let aggregateNSSymbolTable = this.namespaceLookup.get(nameSoFar.toLowerCase()).symbolTable;
                 this.linkSymbolTableDisposables.push(
                     currentNSType.memberTable.addSibling(aggregateNSSymbolTable)
@@ -1291,11 +1303,15 @@ export class Scope {
 interface NamespaceContainer {
     file: BscFile;
     fullName: string;
+    fullNameLower: string;
+    parentNameLower: string;
+    nameParts: Identifier[];
     nameRange: Range;
     lastPartName: string;
+    lastPartNameLower: string;
     statements: Statement[];
-    classStatements: Record<string, ClassStatement>;
-    functionStatements: Record<string, FunctionStatement>;
+    classStatements: Map<string, ClassStatement>;
+    functionStatements: Map<string, FunctionStatement>;
     enumStatements: Map<string, EnumStatement>;
     constStatements: Map<string, ConstStatement>;
     namespaces: Map<string, NamespaceContainer>;