… benchmarks/package.json
… benchmarks/target-runner.ts
… benchmarks/targets/lex-parse-validate.ts
… benchmarks/targets/lex.ts
… benchmarks/targets/parse-brs.ts
… benchmarks/targets/parse-xml.ts
… benchmarks/targets/parse.ts
… benchmarks/targets/transpile-brs.ts
… benchmarks/targets/transpile-xml.ts
… benchmarks/targets/transpile.ts
… benchmarks/targets/validate.ts
… src/Scope.spec.ts
… src/bscPlugin/validation/ScopeValidator.ts
… src/util.spec.ts
… src/util.ts