Adds many performance fixes to try and recover lost speed due to type tracking.
Overall improvement (to a large proprietary project):


| Version         | Parse           | Validate       | Stage           | Transpile      | Overall |
| ------------- | ------------- | ------------- | ------------- | ------------- | ------------- |
| `v0.65.1` | 574.506ms| 6s349.844ms| 368.379ms| 2s25.701ms| 11s450ms|
|  `release-0.66.0` | 636.11ms|9s630.517ms | 377.610ms | 2s75.362ms |  15s687ms |
| this PR | 575.307ms | 5s776.682ms |241.886ms | 1s834.920ms | 10s403ms |




Here are the various improvements made in this PR:

- optimize `util.getAllGottedGetParts`

  ![image](https://github.com/rokucommunity/brighterscript/assets/2544493/b4fb861d-5368-46b5-8aa2-0fb544a82472)
- optimize `util.getDottedGetPath`

  ![image](https://github.com/rokucommunity/brighterscript/assets/2544493/030928c3-f09d-40ba-8e92-2a31779f9b38)

- optimize `util.getAllDottedGetPartsAsString`
    - with old `getAllDottedGetParts`
      ![image](https://github.com/rokucommunity/brighterscript/assets/2544493/d56be8da-20dd-4e9b-9303-db72f18d42ae)
    - with new `getAllDottedGetParts`
      ![image](https://github.com/rokucommunity/brighterscript/assets/2544493/7277f022-7d58-478d-b89b-dc71dd8736fa)

- optimize `cache.getOrAdd` to call `super.has()` and `super.get()` instead of `this.*` which eliminates some prototype lookups. 
  
  ![image](https://github.com/rokucommunity/brighterscript/assets/2544493/c908ace1-4e62-4e32-b1ca-04a27ec702ce)

- optimize scope.buildNamespaceLookup()

  ![image](https://github.com/rokucommunity/brighterscript/assets/2544493/b5cb9a1a-62fb-4460-be63-b1f8ef5e3adc)