@TwitchBronBron 2023-07-10
> Wow 🤩

I wouldn't get _too_ excited. These are just the raw function benchmark numbers for specific functions. The overall validation speeds still aren't changing yet :( I'm hoping after a bunch of these, it will make a difference.